+++ changelog.markdown
@@ -69,3 +69,4 @@
 * Added file **.travis.yml** to control testing by <https://docs.travis-ci.com>.
 * Added file **.ghci**.
 * Replaced use of module **ToolShed.Defaultable** with **Data.Default**.
+* Reimplemented **RegExDot.RegEx.deconstruct** using record-syntax.